==> 01 Preprocess Students.ipynb <==
from hashlib import md5
from matplotlib import pyplot as plt
import numpy as np
import pandas as pd
%matplotlib inline

files = [
    '2017_year.csv',
    '2018_year.csv',
    '2019_year.csv'
]

data_frames = []
for file in files:
    df_raw = pd.read_csv(
        'data/students/' + file,
        converters={'Campus ID': str}
    )
    df_raw['uid'] = pd.Series(index=df_raw.index, dtype='uint64')
    for j in df_raw.index:
        u = str(df_raw.loc[j, 'Campus ID']).lower().encode('utf-8')
        df_raw.loc[j, 'uid'] = md5(u).hexdigest()
    df_raw = df_raw.dropna(subset=['uid']).set_index('uid')
    df = pd.DataFrame(index=df_raw.index)
    df['level'] = df_raw.Level
    df['program'] = df_raw['Program, Plan, and Subplan']
    df.to_csv('preprocessed/student_' + file)
