==> 01 Preprocess Survey.ipynb <==
from matplotlib import pyplot as plt
import numpy as np
import pandas as pd
%matplotlib inline

files = [
    'SI301 Fall 2017 Feedback Form (Early Teamwork Group).csv',
    'SI301 Fall 2017 Feedback Form (Late Teamwork Group).csv',
    'SI301 Fall 2018 Feedback Form (Early Teamwork Group).csv',
    'SI301 Fall 2018 Feedback Form (Late Teamwork Group).csv',
    'SI301 Fall 2019 Feedback Form (Early Teamwork Group).csv',
    'SI301 Fall 2019 Feedback Form (Late Teamwork Group).csv'
]

def make_likert(df):
    likert = {
        'Strongly Disagree': -2,
        'Disagree': -1,
        'Neutral': 0,
        'Agree': 1,
        'Strongly Agree': 2,
        'N/A': float('nan')
    }
    for c in df.columns:
        for i in df.index:
            try:
                parts = df[c][i].split(';')
            except AttributeError:
                continue
            except KeyError:
                continue
            if len(parts) > 0 and parts[0] in likert.keys():
                try:
                    v = np.nanmean([likert[p.strip()] for p in parts])
                except KeyError:
                    print(parts, df[c][i], i)
                df[c][i] = v
    return df

data_frames = []
for i, file in enumerate(files):
    df = make_likert(pd.read_csv('data/' + file))
    data_frames.append(df)
    df.to_csv('preprocessed/' + file)

df
