==> 02_plot_histograms_all.ipynb <==
from matplotlib import pyplot as plt
import numpy as np
import pandas as pd
import scipy.stats as spstats
%matplotlib inline

files = [
    'SI301 Fall 2017 Feedback Form (Early Teamwork Group).csv',
    'SI301 Fall 2017 Feedback Form (Late Teamwork Group).csv',
    'SI301 Fall 2018 Feedback Form (Early Teamwork Group).csv',
    'SI301 Fall 2018 Feedback Form (Late Teamwork Group).csv'
]

data_frames = []
for i, file in enumerate(files):
    data_frames.append(pd.read_csv('preprocessed/' + file, index_col=0))

df_exp = pd.concat([data_frames[0], data_frames[2]])
df_con = pd.concat([data_frames[1], data_frames[3]])
print(len(df_exp), len(df_exp))

exclude = set([
    'Any other comments?',
    'Timestamp',
    'Which aspects of this course do you like best?',
    'Which aspects of this course do you like least?'])

columns = sorted(
    (set(df_exp.columns) & set(df_con.columns)) - exclude)

columns

for c in columns:
    x1 = df_con[c].dropna()
    x2 = df_exp[c].dropna()
    t, pt = spstats.ttest_ind(x1, x2)
    u, pu = spstats.mannwhitneyu(x1, x2)
    plt.figure()
    plt.title("{} ($p_t$={:0.3f} $p_u$={:0.3f})".format(c, pt, pu))
    data = [
        df_con[c].dropna(),
        df_exp[c].dropna()]
    plt.hist(data, bins=[-2.5, -1.5, -0.5, 0.5, 1.5, 2.5], label=['control', 'treatment'])
    plt.legend()
