==> 03_analyze_grades_all.ipynb <==
from matplotlib import pyplot as plt
import numpy as np
import pandas as pd
import scipy.stats as spstats
%matplotlib inline

grade_files = [
    'preprocessed/grades_2017.csv',
    'preprocessed/grades_2018.csv',
    'preprocessed/grades_2019.csv'
]

df_con_year = []
df_exp_year = []

for i, grade_file in enumerate(grade_files):
    df = pd.read_csv(grade_file)
    df_con_year.append(df[df['treatment'] == 'control'])
    df_exp_year.append(df[df['treatment'] == 'experiment'])

df_exp = pd.concat(df_exp_year).dropna()
df_con = pd.concat(df_con_year).dropna()

# Note: treament occurs after midterm in 2019, so this code should be adjusted accordingly if it is used
data = [
    df_con.midterm,
    df_exp.midterm]
t, p = spstats.ttest_ind(df_con.midterm, df_exp.midterm)
plt.figure()
plt.title("{} (t={:0.2f}, p={:0.3f})".format('midterm', t, p))
plt.hist(data, label=['control', 'treatment'])
plt.legend()

data = [
    df_con.final,
    df_exp.final]
t, p = spstats.ttest_ind(df_con.final, df_exp.final)
plt.figure()
plt.title("{} (t={:0.2f}, p={:0.3f})".format('final', t, p))
plt.hist(data, label=['control', 'treatment'])
plt.legend()

spstats.kruskal(df_con.final, df_exp.final)
