==> 04 Regression.ipynb <==
from matplotlib import pyplot as plt
import numpy as np
import pandas as pd
import scipy.stats as spstats
import statsmodels.api as sm
import statsmodels.formula.api as smf
%matplotlib inline

## Load grades

df_2017 = pd.read_csv('preprocessed/grades_2017.csv', index_col='uid')
df_2018 = pd.read_csv('preprocessed/grades_2018.csv', index_col='uid')
df_2019 = pd.read_csv('preprocessed/grades_2019.csv', index_col='uid')

## Load student info

df_student_2017 = pd.read_csv('preprocessed/student_2017_year.csv', index_col='uid')
df_student_2018 = pd.read_csv('preprocessed/student_2018_year.csv', index_col='uid')
df_student_2019 = pd.read_csv('preprocessed/student_2019_year.csv', index_col='uid')
df_2017 = df_2017.join(df_student_2017)
df_2018 = df_2018.join(df_student_2018)
df_2019 = df_2019.join(df_student_2019)

for df in [df_2017, df_2018, df_2019]:
    df['midterm_lost'] = 100 - df.midterm

### Midterms

df = pd.concat([df_2017, df_2018, df_2019])
df = df.dropna(subset=['midterm'])

model = smf.ols(formula='midterm ~ C(teamwork) + C(level) + C(program) + C(year)', data=df).fit()

model.summary()

### Assignments

This regression uses the difference between average teamwork and average individual assignments as the dependent variable, regardless of whether those were early or late. Early/late teamwork is accounted for using the `teamwork` control variable.

The `team_assignment` independent variable represents the assignment method used for the teamwork assignments.

df = pd.concat([df_2017, df_2018, df_2019])

model = smf.ols(formula='team_minus_individual_grades ~ C(team_assignment) + C(teamwork) + C(level) + C(program) + C(year)', data=df).fit()

model.summary()

### Final

df = pd.concat([df_2017, df_2018, df_2019])
df = df.dropna(subset=['final'])

model = smf.ols(formula='final ~ C(team_assignment) +C(teamwork) + C(level) + C(program) + C(year)', data=df).fit()

model.summary()
